Initial Push
This is the first initial push of the current working build. The Plugin is still made for the 2020.12.9s version of Among Us and thus needs to be updated to the new version soon. At the moment there are 8 roles in the code of which are 7 implemented and tested to a certain degree. (Definitely needs some working on edge cases and balancing).