Refactored whole Role-Button System
Roles don't use the standard KillButtonManager anymore. It now uses the RoleButton-System, which is just another layer added on top of the standard KillButtonmanager. This was necessary, because roles without a target weren't able to use their ability without a target in range before (like Survivor or Vigilante).
This also gives way more control for future roles (Like having 2 buttons), because this system is now way more flexible.